pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "duckduckgo-search",
    "jupyter>=1.1.1",
    "langchain-text-splitters>=1.1.1",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pypdf>=6.9.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "sentence-transformers>=5.3.0",
    "sniffio>=1.3.1",
    "streamlit>=1.55.0",
    "tqdm>=4.67.3",
    "zhipuai>=2.1.5.20250825",
]
